backend/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
dj-database-url==3.0.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
firebase-admin==6.7.0
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
msgpack==1.1.0
neon==0.1.2
oauthlib==3.2.2
packaging==24.2
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.3
pytest==8.3.5
pytest-django==4.11.1
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==6.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uuid==1.30
